Add caching to users with kv

+++ src/services/users/users.service.ts
@@ -1,17 +1,25 @@
 import { getAuthToken } from "@/utils/getAuthToken";
 import { HttpAPI } from "../common/http.service";
 import { UserType } from "@/types/user.types";
+import kv from '../../utils/kvClient'
 
 export class UsersService {
   static async getUsers() {
     try {
+      const cachedUsers = await kv.get('users') as any;
+      if (cachedUsers) {
+        return JSON.parse(cachedUsers);
+      }
+
       const token = getAuthToken();
       const response = await HttpAPI.get(`http://localhost:5000/api/v1/users`, {
         headers: {
           Authorization: `Bearer ${token}`,
         },
       });
-      return response;
+
+      await kv.set('users', JSON.stringify(response.data), { ex: 3600 }); // Cache for 1 hour
+      return response.data;
     } catch (error) {
       throw new Error("Error fetching data");
     }
@@ -19,7 +27,6 @@ export class UsersService {
 
   static async addUser(data: UserType) {
     const token = getAuthToken();
-
     if (!token) throw new Error("Token not found");
 
     const response = await HttpAPI.post(
@@ -27,38 +34,45 @@ export class UsersService {
       data,
       token
     );
+
+    await kv.del('users'); 
     return response;
   }
 
   static async deleteUser(email: string) {
     const token = getAuthToken();
 
-    //get the id of the user with the email from the user context
-    const users = await UsersService.getUsers();
-    const id = users.data.find((user: UserType) => user.email === email)?.id;
+    const users = await this.getUsers();
+    const user = users.find((user: UserType) => user.email === email);
+    if (!user) throw new Error("User not found");
 
     if (!token) throw new Error("Token not found");
 
     const response = await HttpAPI.delete(
-      `http://localhost:5000/api/v1/users/${id}`,
+      `http://localhost:5000/api/v1/users/${user.id}`,
       token
     );
+
+    await kv.del('users'); // Invalidate cache
     return response;
   }
 
   static async updateUser(data: UserType) {
     const token = getAuthToken();
 
-    const users = await UsersService.getUsers();
-    const id = users.data.find((user: UserType) => user.id === data.id)?.id;
+    const users = await this.getUsers();
+    const user = users.find((user: UserType) => user.id === data.id);
+    if (!user) throw new Error("User not found");
 
     if (!token) throw new Error("Token not found");
 
     const response = await HttpAPI.patch(
-      `http://localhost:5000/api/v1/users/update/${id}`,
+      `http://localhost:5000/api/v1/users/update/${user.id}`,
       data,
       token
     );
+
+    await kv.del('users'); // Invalidate cache
     return response;
   }
 }
